pyproject.toml
[tool.poetry]
name = "pdf-signer-desktop"
version = "0.1.0"
description = "Aplicación de firmado de pdf's"
authors = ["eder3232 <eldebar.dk@gmail.com>"]
readme = "README.md"
packages = [{include = "app"}]

[tool.poetry.dependencies]
python = ">=3.12,<3.13"
PyPDF2 = "^3.0.0"
Pillow = "^10.0.0"
pydantic = "^2.0.0"
PySide6 = "^6.0.0"
reportlab = "^4.0.0"
pymupdf = "^1.26.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
app = "main:main"
